scrapers/social_media_scraper.py
"""
Social Media Data Scraper
Collects follower counts and engagement metrics for brand valuation
"""

import requests
from bs4 import BeautifulSoup
import re
import time
import logging
from typing import Dict, Optional
from pathlib import Path
import sys

# Add project root to path
project_root = Path(__file__).parent.parent
sys.path.insert(0, str(project_root))

from config import get_config

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class SocialMediaScraper:
    """
    Scraper for social media metrics
    Note: For production, consider using official APIs when available
    """
    
    def __init__(self):
        self.config = get_config()
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': self.config.get('scraping.user_agent', 'Mozilla/5.0')
        })
        self.request_delay = self.config.get('scraping.request_delay', 2)
    
    def _wait(self):
        """Respectful delay between requests"""
        time.sleep(self.request_delay)
    
    def search_instagram_profile(self, player_name: str, team_name: str = None) -> Optional[Dict]:
        """
        Search for Instagram profile information
        
        Note: This is a simplified version. For production:
        - Use Instagram's official API
        - Implement proper authentication
        - Handle rate limits
        
        Args:
            player_name: Player's name
            team_name: Optional team name for context
        
        Returns:
            Dictionary with Instagram data or None
        """
        logger.info(f"Searching Instagram for {player_name}")
        
        # This is a placeholder - actual implementation would need:
        # 1. Instagram Graph API access
        # 2. Proper authentication
        # 3. Handle of lookup database
        
        return {
            'handle': None,
            'followers': None,
            'engagement_rate': None,
            'verified': False,
            'note': 'Manual lookup required - add to database directly'
        }
    
    def search_twitter_profile(self, player_name: str, team_name: str = None) -> Optional[Dict]:
        """
        Search for Twitter/X profile information
        
        Note: Requires Twitter API access for production use
        
        Args:
            player_name: Player's name
            team_name: Optional team name for context
        
        Returns:
            Dictionary with Twitter data or None
        """
        logger.info(f"Searching Twitter for {player_name}")
        
        # Placeholder - actual implementation needs Twitter API v2
        return {
            'handle': None,
            'followers': None,
            'engagement_rate': None,
            'verified': False,
            'note': 'Manual lookup required - add to database directly'
        }
    
    def estimate_brand_score(self, social_data: Dict) -> float:
        """
        Calculate brand score from social media metrics
        
        Args:
            social_data: Dictionary with social media metrics
        
        Returns:
            Brand score (0-100)
        """
        score = 0.0
        
        # Follower counts (60% of score)
        total_followers = (
            (social_data.get('twitter_followers') or 0) +
            (social_data.get('instagram_followers') or 0) +
            (social_data.get('tiktok_followers') or 0)
        )
        
        # Logarithmic scale for followers
        if total_followers > 0:
            import math
            follower_score = min(60, math.log10(total_followers + 1) * 10)
            score += follower_score
        
        # Engagement rate (20% of score)
        avg_engagement = social_data.get('average_engagement_rate', 0)
        if avg_engagement:
            engagement_score = min(20, avg_engagement * 400)  # 5% engagement = 20 points
            score += engagement_score
        
        # Growth rate (10% of score)
        growth_rate = social_data.get('follower_growth_rate_30d', 0)
        if growth_rate:
            growth_score = min(10, growth_rate * 100)  # 10% growth = 10 points
            score += growth_score
        
        # Media coverage (10% of score)
        if social_data.get('national_media_coverage'):
            score += 10
        
        return min(100, score)
    
    def get_google_trends_data(self, player_name: str) -> Optional[int]:
        """
        Estimate Google search volume (simplified)
        
        For production, integrate with:
        - Google Trends API
        - Google Ads Keyword Planner
        
        Args:
            player_name: Player's name
        
        Returns:
            Estimated monthly search volume or None
        """
        logger.info(f"Checking Google trends for {player_name}")
        
        # Placeholder - would need Google Trends API
        return None


class NILDatabaseScraper:
    """
    Scraper for public NIL deal information
    Sources: On3, 247Sports, Opendorse, etc.
    """
    
    def __init__(self):
        self.config = get_config()
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': self.config.get('scraping.user_agent', 'Mozilla/5.0')
        })
    
    def search_on3_nil_deals(self, player_name: str) -> list:
        """
        Search On3 NIL database for player deals
        
        Note: This is a placeholder. Actual implementation would need:
        - Proper web scraping with BeautifulSoup
        - Respect for robots.txt
        - Rate limiting
        
        Args:
            player_name: Player's name
        
        Returns:
            List of NIL deals
        """
        logger.info(f"Searching On3 NIL database for {player_name}")
        
        # Placeholder - actual implementation needed
        return []
    
    def get_nil_valuation_estimate(self, player_data: Dict) -> float:
        """
        Estimate NIL value based on available data
        
        Args:
            player_data: Player information including social media
        
        Returns:
            Estimated annual NIL value
        """
        # Basic estimation formula
        base_value = 0
        
        # Position multiplier
        position_multipliers = {
            'QB': 2.5,
            'WR': 1.5,
            'RB': 1.4,
            'TE': 1.1,
            'DL': 1.3,
            'LB': 1.2,
            'CB': 1.2,
            'S': 1.1,
            'OL': 0.9
        }
        
        position = player_data.get('position', 'Unknown')
        multiplier = position_multipliers.get(position, 1.0)
        
        # Social media followers
        total_followers = (
            player_data.get('instagram_followers', 0) +
            player_data.get('twitter_followers', 0) +
            player_data.get('tiktok_followers', 0)
        )
        
        # Rough estimation: $1-5 per follower annually for college athletes
        if total_followers > 100000:
            follower_value = total_followers * 3
        elif total_followers > 50000:
            follower_value = total_followers * 2
        elif total_followers > 10000:
            follower_value = total_followers * 1
        else:
            follower_value = total_followers * 0.5
        
        base_value = follower_value * multiplier
        
        # Performance bonus (if available)
        if player_data.get('is_starter'):
            base_value *= 1.5
        
        # Team prestige bonus
        power_5_teams = ['Alabama', 'Ohio State', 'Georgia', 'Texas', 'USC', 
                        'Michigan', 'LSU', 'Florida', 'Oregon', 'Penn State']
        
        if player_data.get('team') in power_5_teams:
            base_value *= 1.3
        
        return base_value


# Example usage
if __name__ == '__main__':
    scraper = SocialMediaScraper()
    
    # Example player data
    player = {
        'name': 'Sample Player',
        'position': 'QB',
        'team': 'Alabama',
        'instagram_followers': 150000,
        'twitter_followers': 85000,
        'average_engagement_rate': 0.045,
        'national_media_coverage': True,
        'is_starter': True
    }
    
    brand_score = scraper.estimate_brand_score(player)
    print(f"Brand Score: {brand_score:.1f}")
    
    nil_scraper = NILDatabaseScraper()
    nil_estimate = nil_scraper.get_nil_valuation_estimate(player)
    print(f"Estimated NIL Value: ${nil_estimate:,.0f}/year")

