requirements.txt
# College Athlete Valuation Model - Dependencies

# Core numerical computing
numpy>=1.24.0
pandas>=2.0.0

# Database
sqlalchemy>=2.0.0          # ORM for database operations
alembic>=1.12.0           # Database migrations

# API and Web
requests>=2.31.0          # HTTP requests
pyyaml>=6.0.1            # Configuration files
beautifulsoup4>=4.12.0   # Web scraping
lxml>=4.9.0              # HTML/XML parser
selenium>=4.15.0         # Dynamic content scraping (optional)

# Rate limiting and caching
ratelimit>=2.2.1         # API rate limiting
requests-cache>=1.1.0    # Response caching

# Optional but recommended
matplotlib>=3.7.0         # For visualizations
scipy>=1.10.0            # For statistical analysis

# Development tools (optional)
pytest>=7.4.0            # For testing
jupyter>=1.0.0           # For notebooks

# Dashboard and visualization
streamlit>=1.28.0        # Web dashboard framework
plotly>=5.17.0          # Interactive charts
altair>=5.1.0           # Declarative visualizations
